Fix indentation handling
The previous implementation did not support tab indentations or deeply
nested types. Instead of using indentationWidth to pad the generated
code, we deduce the current indentation depth from the first line in the
selection and check if the user prefers tabs for the initializer body.

This commit also inserts a new line between the variables and generated
initializers, as this seems to be a common Swift design choice.